requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
chromadb==0.5.5
pandas==2.2.2
sentence_transformers
python-dotenv==1.0.1
jinja2==3.1.4
httpx==0.27.2
pytest==8.3.2
black==24.8.0
ruff==0.6.9
requests==2.32.3
prometheus-client==0.20.0
reportlab==4.2.2
Jinja2==3.1.4
httpx==0.27.2
pyjwt